refactor(#343): cache module option 동적으로 할당 및 Ttl, redis service 경로 변경
